+++ src/utils/utf8-to-jis-table.js
@@ -16,8 +16,8 @@ const b64Int = x => xInt(base64ToHex(x))
 function decompressUtf8ToJisTable (compressedTable) {
   /** @type {Record<number, number>} */
   const result = {}
-  for (const [jisChar, utf8ValsStr] of Object.entries(compressedTable)) {
-    const utf8Vals = utf8ValsStr.split(',')
+  for (const [jisChar, compressedUtf8ValsStr] of Object.entries(compressedTable)) {
+    const utf8Vals = decompressUtf8ValsStr(compressedUtf8ValsStr).split(',')
     let charIterator = xInt(jisChar)
     for (const utf8Value of utf8Vals) {
       if (utf8Value.includes(':')) {
@@ -34,6 +34,35 @@ function decompressUtf8ToJisTable (compressedTable) {
       }
     }
   }
+  return result
+}
+
+/**
+ * decompress each value in utf8-to-jis-table.constants.js object (e.g "4[p[S[A,C,M,Q,Y,U,c,s,k,0,8,B,D,P,T,b,X,j,z,r,7],WL,S[g,v,o,3,/,d,w,l,4],WC]]" to
+ * "4pSA,4pSC,4pSM,4pSQ,4pSY,4pSU,4pSc,4pSs,4pSk,4pS0,4pS8,4pSB,4pSD,4pSP,4pST,4pSb,4pSX,4pSj,4pSz,4pSr,4pS7,4pWL,4pSg,4pSv,4pSo,4pS3,4pS/,4pSd,4pSw,4pSl,4pS4,4pWC")
+ * @param {string} compressedVals - compressed utf8-to-jis-table.constants.js field value
+ */
+function decompressUtf8ValsStr (compressedVals) {
+  let result = ''
+  let prefix = ''
+  let initPrefix = ''
+  for (const char of compressedVals) {
+    switch (char) {
+      case '[':
+        initPrefix = prefix
+        break
+      case ']':
+        initPrefix = initPrefix.slice(0, -1)
+        break
+      case ',':
+        result += prefix + ','
+        prefix = initPrefix
+        break
+      default:
+        prefix += char
+    }
+  }
+  result += prefix
   return result
 }